+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # CppJieba ChangeLog
 
-## next version
+## v4.5.0
 
 + 在 Trie 中去除之前糟糕的针对 uint16 优化的用数组代替 map 的设计，
 该设计的主要问题是前提 unicode 每个字符必须是 uint16 ，则无法更全面得支持 unicode 多国字符。